pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vibelint"
version = "0.1.0"
description = "AI code pattern scanner — catch the bugs AI writes 2.74x more often"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Tanmay Shrivastava" }]
keywords = ["ai", "linter", "static-analysis", "vibe-coding", "tree-sitter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Quality Assurance",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "tree-sitter>=0.22",
    "tree-sitter-typescript>=0.23",
    "tree-sitter-javascript>=0.23",
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[project.scripts]
vibelint = "vibelint.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
